Move all code to lib/texas_holdem
