+++ lib/texas_holdem/bet_manager.rb
