+++ lib/texas_holdem/card.rb
