+++ lib/texas_holdem/cards.rb
