+++ lib/texas_holdem/hand.rb
