+++ lib/texas_holdem/holdem_bet_manager.rb
@@ -1,4 +1,4 @@
-require_relative '../bet_manager.rb'
+require_relative './bet_manager.rb'
 
 class HoldemBetManager < BetManager
   def bet(players)
