+++ lib/texas_holdem/player.rb
